+++ src/plone/restapi/tests/test_dxcontent_serializer.py
@@ -7,10 +7,13 @@ from datetime import timedelta
 from plone.app.testing import TEST_USER_ID
 from plone.app.testing import setRoles
 from plone.restapi.interfaces import IExpandableElement
+from plone.app.textfield.interfaces import ITransformer
+from plone.app.textfield.value import RichTextValue
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.testing import PLONE_RESTAPI_DX_INTEGRATION_TESTING
 from plone.restapi.tests.test_expansion import ExpandableElementFoo
 from plone.uuid.interfaces import IMutableUUID
+from zope.component import getGlobalSiteManager
 from zope.component import getMultiAdapter
 from zope.component import provideAdapter
 from zope.interface import Interface
@@ -20,6 +23,22 @@ import json
 import unittest
 
 
+class AdapterCM(object):
+    """Context manager that will temporarily register an adapter
+    """
+    def __init__(self, adapter, from_, provides):
+        self.adapter = adapter
+        self.from_ = from_
+        self.provides = provides
+        self.gsm = getGlobalSiteManager()
+
+    def __enter__(self):
+        self.gsm.registerAdapter(self.adapter, self.from_, self.provides)
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        self.gsm.unregisterAdapter(self.adapter, self.from_, self.provides)
+
+
 class TestDXContentSerializer(unittest.TestCase):
 
     layer = PLONE_RESTAPI_DX_INTEGRATION_TESTING
@@ -27,6 +46,12 @@ class TestDXContentSerializer(unittest.TestCase):
     def setUp(self):
         self.portal = self.layer['portal']
         self.request = self.layer['request']
+
+        richtext_value = RichTextValue(
+            u'Käfer',
+            'text/plain',
+            'text/html'
+        )
 
         self.portal.invokeFactory(
             'DXTestDocument',
@@ -47,6 +72,7 @@ class TestDXContentSerializer(unittest.TestCase):
             test_list_field=[1, 'two', 3],
             test_set_field=set(['a', 'b', 'c']),
             test_text_field=u'Käfer',
+            test_richtext_field=richtext_value,
             test_textline_field=u'Käfer',
             test_time_field=time(14, 15, 33),
             test_timedelta_field=timedelta(44),
@@ -151,3 +177,28 @@ class TestDXContentSerializer(unittest.TestCase):
         obj = serializer()
         self.assertIn('is_folderish', obj)
         self.assertEquals(True, obj['is_folderish'])
+
+    def test_richtext_serializer_context(self):
+        """This checks if the context is passed in correctly.
+
+           We define a ITransformer, which returns the contexts portal_type.
+           This is then verfied.
+        """
+
+        class RichtextTransform(object):
+            """RichttextValue to show that the context is correctly passed
+               in throughout the stack.
+            """
+            def __init__(self, context):
+                self.context = context
+
+            def __call__(self, value, mime_type):
+                return self.context.portal_type
+
+        with AdapterCM(RichtextTransform, (Interface, ), ITransformer):
+            obj = self.serialize()
+
+        self.assertEquals(
+            obj['test_richtext_field']['data'],
+            self.portal.doc1.portal_type
+        )